Extract LayersSwitcher and remove duplicated functions
